Station rewards in different file
